requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
Pillow>=10.0.0
reportlab>=4.0.0
langgraph>=0.1.0
langchain>=0.1.0
langchain-openai>=0.1.0
tiktoken>=0.5.0
kaggle>=1.5.0
fastapi>=0.104.0
uvicorn>=0.24.0
joblib>=1.3.0
mlflow>=2.8.0
pytest>=7.0.0
flake8>=6.0.0
openai>=1.0.0
requests>=2.31.0